refactored event creation
