+++ lib/util/schema/event.js
@@ -7,8 +7,7 @@ module.exports = Joi.object().keys({
   location: Joi.object().keys({
     name: Joi.string().min(1).required(),
     lat: Joi.number().min(-90).max(90).required(),
-    lng: Joi.number().min(-180).max(180).required(),
-    timeZone: Joi.string().valid(require('moment-timezone').tz.names()).required()
+    lng: Joi.number().min(-180).max(180).required()
   }).required(),
   startDate: date.required(),
   startTime: time,
